# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/room-types": {
      "get": {
        "operationId": "RoomTypesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      },
      "post": {
        "operationId": "RoomTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      }
    },
    "/v1/room-types/{id}": {
      "get": {
        "operationId": "RoomTypesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      },
      "patch": {
        "operationId": "RoomTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "ETag obtenu à la lecture",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      }
    },
    "/v1/rooms": {
      "get": {
        "operationId": "RoomsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Jeton `nextCursor` de la page précédente",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Champs à renvoyer, séparés par des virgules (ex. id,status,guest)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "available",
                "occupied",
                "cleaning",
                "maintenance"
              ]
            }
          },
          {
            "name": "room_type_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "number | -number | floor | -floor",
            "schema": {
              "example": "number",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      }
    },
    "/v1/rooms/{id}": {
      "get": {
        "operationId": "RoomsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "ETag obtenu à la lecture",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chambres"
        ]
      }
    },
    "/v1/rooms/{id}/status": {
      "patch": {
        "operationId": "RoomsController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "ETag obtenu à la lecture",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer l’état d’une chambre (propre, nettoyage, maintenance)",
        "tags": [
          "Chambres"
        ]
      }
    },
...
  },
  "info": {
    "title": "API – Système de gestion hôtelière",
    "description": "Socle V1 : session, menu dynamique, chambres, clients, disponibilités, réservations, réception.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRoomTypeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "DBL"
          },
          "name": {
            "type": "string",
            "example": "Chambre double"
          },
          "capacity": {
            "type": "number",
            "example": 2
          },
          "basePrice": {
            "type": "number",
            "example": 9500,
            "description": "Tarif de base par nuit en unités mineures"
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          }
        },
        "required": [
          "code",
          "name",
          "capacity",
          "basePrice"
        ]
      },
      "UpdateRoomTypeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "DBL"
          },
          "name": {
            "type": "string",
            "example": "Chambre double"
          },
          "capacity": {
            "type": "number",
            "example": 2
          },
          "basePrice": {
            "type": "number",
            "example": 9500,
            "description": "Tarif de base par nuit en unités mineures"
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          }
        }
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string",
            "example": "204"
          },
          "floor": {
            "type": "number",
            "example": 2
          },
          "roomTypeId": {
            "type": "string"
          }
        },
        "required": [
          "number",
          "roomTypeId"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string",
            "example": "204"
          },
          "floor": {
            "type": "number",
            "example": 2
          },
          "roomTypeId": {
            "type": "string"
          }
        }
      },
      "RoomStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "available",
              "occupied",
              "cleaning",
              "maintenance"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}